Move SSH and related to netutil
