+++ lib/installer.py
@@ -2,8 +2,8 @@ import logging
 import time
 import xml.etree.ElementTree as ET
 
-from lib.commands import SSHCommandFailed, ssh
 from lib.common import wait_for
+from lib.netutil import SSHCommandFailed, ssh
 
 class AnswerFile:
     def __init__(self, kind, /):
