+++ tests/migration/test_host_evacuate.py
@@ -2,8 +2,8 @@ import pytest
 
 import logging
 
-from lib.commands import SSHCommandFailed
 from lib.common import wait_for
+from lib.netutil import SSHCommandFailed
 
 # The pool needs a shared SR to use `host.evacuate`. All three fixtures below are needed.
 from tests.storage.nfs.conftest import nfs_device_config, nfs_sr, vm_on_nfs_sr
