+++ tests/storage/xfs/test_xfs_sr.py
@@ -5,8 +5,8 @@ import pytest
 import logging
 import time
 
-from lib.commands import SSHCommandFailed
 from lib.common import vm_image, wait_for
+from lib.netutil import SSHCommandFailed
 from tests.storage import vdi_is_open
 
 from typing import TYPE_CHECKING
