+++ tests/storage/zfs/test_zfs_sr.py
@@ -3,8 +3,8 @@ import pytest
 import logging
 import time
 
-from lib.commands import SSHCommandFailed
 from lib.common import vm_image, wait_for
+from lib.netutil import SSHCommandFailed
 from tests.storage import vdi_is_open
 
 from .conftest import POOL_NAME, POOL_PATH
