+++ tests/uefi_sb/test_auth_var.py
@@ -1,12 +1,12 @@
 import pytest
 
-from lib.commands import SSHCommandFailed
 from lib.efi import (
     EFI_AT_ATTRS_BYTES,
     Certificate,
     EFIAuth,
     global_variable_guid,
 )
+from lib.netutil import SSHCommandFailed
 
 # Requirements:
 # On the test runner:
