+++ tests/xen/test_xtf.py
@@ -2,7 +2,7 @@ import pytest
 
 import logging
 
-from lib.commands import SSHCommandFailed
+from lib.netutil import SSHCommandFailed
 
 # Requirements:
 # From --hosts parameter:
